post-interaction-reel-service/requirements.txt
# =============================================================================
# POST, INTERACTION & REEL SERVICE - REQUIREMENTS.TXT
# =============================================================================
# Các thư viện cần thiết để chạy dự án

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Redis (optional - có thể bỏ nếu không dùng)
redis==5.0.1

# File Storage (optional - có thể bỏ nếu không dùng)
boto3==1.34.0
minio==7.2.0

# Video Processing (optional - có thể bỏ nếu không dùng)
ffmpeg-python==0.2.0
pillow==10.1.0

# Notifications (optional - có thể bỏ nếu không dùng)
firebase-admin==6.4.0

# HTTP Client
httpx==0.25.2

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
